cgame/game: rework grenade tick handle and remove dynamite from it
Dynamite tick is not use anymore, old relic from SP when it was possible to set the timer on the dynamite and each tick was increasing the timer by 5 sec.

Dynamite don't use grenade time left anymore. But the old behaviour (waiting with dynamite in hand and the tick sound) are preserved (nostalgia).

The grenade time left value is now in table weapon and the animation for all throwable explosive weapon are merge in the same condition.