cgame: merge shoutcast minimap with compass and allow usage of square minimap, refs #1000 #1282 #1433
Initialy I separated shoutcast minimap and compass because I didn't want to break compass as I barely knew the project. I did it however in mind of eventually adding it back and allowing using it for normal gameplay and this PR should be it.

    standard compass should be working as before
    shoutcast minimap should be working as before - it is still forced in both the size and placement
    enabled with cg_drawCompass 2

I would consider it a work in progress, standard cg_altHud settings will correctly draw the square minimap, but it is hardly usable. It's too small and icons are too big for such small size. As of now it is best used with custom hud, where players can set the size and location as they wish.

What I feel could and probably should be worked on in the future:

    improve the icons scaling - square minimap uses big icons so they are easily visible when max zoomed out but with small minimap dimensions it barely presents any information because of it. Compass on the other hand has very small icons when max zoomed out making it only usable with big zoom which reduces the overall needed "visibility".
    fix voice chat sprites - on square minimap they can be drawn out side of minimap (same as compass but it has bigger border). Position of sprites is also pretty bad - full zoom it is under the player icon, full zoom out it is way too far from player icon.
    to make square minimap work with current cg_altHuds it is also reduced in size just like compass which shouldn't be necessary imo. https://github.com/etlegacy/etlegacy/compare/master...ryzyk-krzysiek:square-minimap?expand=1#diff-09bf4aee5c64caee30f85eb55b903a2b2004f3461eb16040b223a25ff18523b8R1771-R1775

Custom hud example:
square minimap

refs #1000 #1282 #1433