mod: add missile camera for practice, refs #1898
only enabled with cheats. Competition servers can add practice config where cheats are enabled. It would be nice, if it was enabled also in warmup before competitive matches:
Allow cheats in warmup for practise purposes #1898
position, size and visibility configurable by custom HUD - @rmarquis mentioned it's not good to have it always visible when cheats are enabled
missile entity is not unlinked when going through skybox, so player can still see camera in certain cases. This makes missile camera always visible on supply, but not on goldrush where skybox is really low. Would be nice if someone could also fix it for goldrush.
camera won't show entities that are not in his PVS like in NQ, In fact it doesn't show any entities, but I don't know why. I don't think its needed for practice