server: sv_userInfoFloodProtect to protect against userinfo flooding, refs #1906
After the changes made in #1906, sv_floodProtect is mostly forced off on legacy servers, since it interferes with mod-sided flood protection. However userinfo flooding was tied to this cvar in 25b5628, so it was no longer being protected against. This adds a new cvar sv_userInfoFloodProtect which protects specifically against that, so sv_floodProtect can be off and server can still be protected against user info flooding.