client: reduces latency, improves debugging output, creates error correction flags (#2043)
Two bit flags were added to the user command packet to notify the server that the client sense of time was adjusted. This allows for special handling of serverTime in the next command received. Temporary debugging output was added to mod code in bg_pmove.c with g_developer 2. This change gives legacy and other mods the information needed to properly respond to incorrect serverTime values caused by automatic adjustments to serverTimeDelta on the client.

Creates cl_extrapolationMargin with a default value of 1 and allowed range 0 to 10 to replace a magic number 5 that added unnecessary latency at some client frame rates. This change will typically reduce the client latency by up to 4 ms at the default value. Servers that want to restrict client control can use sv_cvar to force default.

When the client sense of time in incremented is now more cautious to accommodate this change instead of always moving forward every snapshot.