game: fix explosion hit registration being inconsistent
As far as I know the order of entities returned by trap_EntitiesInBox is linking order. This PR aim is to make the player entities in the list be sorted by distance from the explosion origin in ascending order. It would fix situations where player A can block damage to player B, but player B cannot block damage to player A.

This doesn't change how players can be blocking damage by itself.