mod: scale text shadow/outline based off text size
Use scaling relative to text size rather than hard-coded 1/2px offset for drawing shadows and outline on text elements. This greatly improves readability especially on small text sizes, because the amount of shadowing now shrinks together with the text size, keeping the "ratio" of text/shadow the same. Bigger text on the other hand becomes slightly less readable when going to big sizes, but this is only really going to start to matter on text sizes which are so big that nobody is likely going to use.

The initial offset that gets scaled based off text size is just an arbitrary value picked by me, it can definitely be changed if the proposed value produces too little or too much shadowing.