client: switched to using json with the favcache data
The old format that just dumps the structs to drive as is, was broken
many times due to added fields on the struct and no actual processing on
the in and output of the file. Essentially the old file could not be
trusted, but the file was not really used anymore except for cases where
the database was disabled.