mod: remove WBUTTON_ATTACK2 (+attack2)
It seems to me that WBUTTON_ATTACK2 should mirror BUTTON_ATTACK inside pmove but is not. This makes current implementation inconsistent, buggy and there are some exploits because of that. I don't see a point of making +attack2 exactly same as +attack though if there is nothing in the game using it anyway. I guess it was added as alternative fire but there also is no such thing in weapontable so both attacks use same values currently but +attack2 lacks important code to behave exactly as +attack.