cgame: fix kick angles framerate dependency, refs #2264
Reduced the step so the kick angles are properly split up into smaller chunks. This isn't perfect as it means everyone will play now with kick angles as if they are playing on 333 fps, also it assumes that 333fps is max possible fps and going above it will result in smaller kick angles again. This is the easiest way to fix the issue without reworking everything most likely.