mod: remove unnecessary trace for water hitmark detection
Removed additional G_Trace from Bullet_Fire_Extended that was tracing for water for some better visual "hitmark" when player missed. That trace meant that for every missed shot server would re-do all head hitboxes for everyone on server when it was already determined that it hit noone. That additional trace for water is also already done client side too as far as I know so it is pure waste of time.