cgame: fix allied rank icons using wrong shaders, fixes #1251, #2615
Because the shader file in legacy pk3 does not match the name of the shader in pak0.pk3, the game does not reliably override the shaders with same names from the pak0's gfx_hud.shader. Using unique names that are only present in the shaders used in legacy pk3, we ensure we're using the correct shaders.