mod: properly hide mouse cursor if menu bypasses mouse input, fix #724
Instead of doing silly hacks that move the mouse cursor off-screen when opening a menu which sets cl_bypassmouseinput 1, just don't draw it in the first place. This also fixes the cursor position resetting when opening a menu which bypasses mouse input, retaining the previous position when a mouse-controlled menu was opened the last time.

The cursor drawing is a somewhat complex system so I'd appreciate if some testing was done on this, to make sure the cursor is visible/hidden in the correct places. I've tested the following menus:

    main/in-game menu
    vsay menu
    ft menu
    class/teams/spawnpoint menu
    chat
    limbo
    speaker editor
    HUD editor
    demo info window (pause/seek/rewind controls)
    intermission
    shoutcaster mode (since it has it's own event handling, cursor isn't drawn)