game: Fixup jump-crouch- and swim-hitbox (#2834)
'ClientHitboxMaxZ' and 'CG_ClientHitboxMaxZ' both are used to determine
the hitbox of a player server and client side respectively.

In the case of holding crouch while in the air jumping,
'CG_ClientHitboxMaxZ' erroneously set the hitbox height to shoulder
level. This is corrected by checking for not being in the air when the
crouch flag is set.

In the case of swimming, crouching while swimming would also set the
hitbox height to shoulder level.

Instead of removing this, it is made the default - as the swimming
hitbox was comically tall anyhow and it is not expected to have a huge
impact towards gameplay.