game: Fix 0-velocity missile bouncing scenarios (#3044)
In some scenarios missiles can trigger 'G_MissileImpact' and thus
'G_BounceMissile' over and over again, despite a bounce velocity of
{0,0,0}.

This causes the entity to rotate and play the 'missileBouncingSound'
over and over again, without actually breaking out of the loop.

This is addressed by exiting early when 'G_BounceMissile' calculates a
velocity of {0,0,0} and the surrounding callsite in 'G_MissileImpact' to
skip sending the 'EV_GRENADE_BOUNCE' event, which in turn would play
'missileBouncingSound' on the clients.

One way to reproduce:

1. Pick engineer
2. Give yourself Engineer 4 Flak Jacket Skill
3. Throw a landmine to your feet (but don't arm it)
4. Switch and throw/prime a grenade to your feet
5. Immediately switch to pliers and arm the landmine
6. The grenade will explode and down you. Immediately exit to limbo.
7. The landmine should now collide with your dead body and be stuck in
   the above missile bounce loop scenario until the body disappears
   (sinks into the ground).